Implement create_txn_cmd
Implemented tauri command to create transactions (txns). Learned
about the unit type () and about error conversion (to String) that
is necessary so tauri can send errors to the front end.

Still not sure how the arguments of tauri commands are mapped to arguments
in typescript land. Also confused about where we need ? and OK().